Move code to get the example data bundle to a bot
