+++ examples/medplum-scheduling-demo/src/scripts/deploy-bots.ts
@@ -27,6 +27,10 @@ const Bots: BotDescription[] = [
     src: 'src/bots/core/block-availability.ts',
     dist: 'dist/bots/core/block-availability.js',
   },
+  {
+    src: 'src/bots/example/example-data.ts',
+    dist: 'dist/bots/example/example-data.js',
+  },
 ];
 
 async function main(): Promise<void> {
